+++ experimental/execution/ptx-executor/src/lib.rs
@@ -8,16 +8,18 @@
 
 mod analyzer;
 mod common;
-mod executor;
 mod finalizer;
 mod metrics;
-mod scheduler;
+mod runner;
+mod sorter;
 mod state_reader;
 mod state_view;
 
+mod scheduler;
+
 use crate::{
-    analyzer::PtxAnalyzer, executor::PtxExecutor, finalizer::PtxFinalizer, metrics::TIMER,
-    scheduler::PtxScheduler, state_reader::PtxStateReader,
+    analyzer::PtxAnalyzer, finalizer::PtxFinalizer, metrics::TIMER, runner::PtxRunner,
+    scheduler::PtxScheduler, sorter::PtxSorter, state_reader::PtxStateReader,
 };
 use aptos_executor::{
     block_executor::TransactionBlockExecutor, components::chunk_output::ChunkOutput,
@@ -47,34 +49,44 @@ impl VMExecutor for PtxBlockExecutor {
         _maybe_block_gas_limit: Option<u64>,
     ) -> Result<Vec<TransactionOutput>, VMStatus> {
         let _timer = TIMER.timer_with(&["block_total"]);
-        // 1. Analyze: annotate read / write sets.
-        // 2. Schedule: build dependency graph by remembering the latests writes for each key.
-        // 3. Execute: start executing a transaction once its dependencies are met.
-        //             and inform dependent transactions after execution.
-        // 5. Finalize: materialize aggregators, etc.
+
         let concurrency_level = AptosVM::get_concurrency_level();
+        // 1. Analyze: annotate read / write sets.
+        // 2. Sort: build dependency graph by remembering the latest writes for each key.
+        // 3. Schedule: send readily runnable transactions to the runner.
+        // 4. Run: in a pool of workers, inform txn outputs to the scheduler to unblock others.
+        // 5. Finalize: materialize aggregators.
+        // And, there is the state reader that asynchronously does the DB reads for the scheduler.
+        // -- in total we need 6 threads other than the runner worker.
         assert!(
-            concurrency_level > 5,
+            concurrency_level > 6,
             "Each of the components needs its own main thread."
         );
-        let num_executor_workers = concurrency_level - 5;
+        let num_executor_workers = concurrency_level - 6;
 
         let ret = Arc::new(Mutex::new(None));
         let ret_clone = ret.clone();
         RAYON_EXEC_POOL.scope(move |scope| {
             let num_txns = transactions.len();
             let (result_tx, result_rx) = channel();
-            let finalizer = PtxFinalizer::spawn(scope, state_view, result_tx);
-            let executor = PtxExecutor::spawn(scope, state_view, num_executor_workers, finalizer);
-            let state_reader = PtxStateReader::spawn(scope, executor.clone(), state_view);
-            let scheduler = PtxScheduler::spawn(scope, executor, state_reader);
-            let analyzer = PtxAnalyzer::spawn(scope, scheduler);
 
+            // Spawn all the components.
+            let finalizer = PtxFinalizer::spawn(scope, state_view, result_tx);
+            let runner = PtxRunner::spawn(scope, state_view, finalizer);
+            let scheduler = PtxScheduler::spawn(scope, runner.clone());
+            runner.spawn_workers(scheduler.clone(), num_executor_workers);
+            let state_reader = PtxStateReader::spawn(scope, scheduler.clone(), state_view);
+            let sorter = PtxSorter::spawn(scope, scheduler, state_reader);
+            let analyzer = PtxAnalyzer::spawn(scope, sorter);
+
+            // Feed the transactions down the pipeline.
             for txn in transactions {
                 analyzer.analyze_transaction(txn);
             }
             analyzer.finish_block();
 
+            // Collect results from the other side of the pipeline and hand over to outside of the
+            // scope.
             let mut txn_outputs = vec![];
             while let Ok(txn_output) = result_rx.recv() {
                 txn_outputs.push(txn_output);
