+++ experimental/execution/ptx-executor/src/runner.rs
@@ -1,0 +1,298 @@
+// Copyright © Aptos Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+#![forbid(unsafe_code)]
+
+//! TODO(aldenhu): doc
+
+use crate::{
+    common::TxnIdx,
+    finalizer::PtxFinalizerClient,
+    metrics::{PER_WORKER_TIMER, TIMER},
+    scheduler::PtxSchedulerClient,
+    state_view::OverlayedStateView,
+};
+use aptos_logger::trace;
+use aptos_metrics_core::TimerHelper;
+use aptos_state_view::StateView;
+use aptos_types::{
+    state_store::{state_key::StateKey, state_value::StateValue},
+    transaction::Transaction,
+    write_set::TransactionWrite,
+};
+use aptos_vm::{
+    adapter_common::{preprocess_transaction, VMAdapter},
+    data_cache::AsMoveResolver,
+    AptosVM,
+};
+use aptos_vm_logging::log_schema::AdapterLogSchema;
+use rayon::Scope;
+use std::{
+    collections::HashMap,
+    sync::mpsc::{channel, Receiver, Sender},
+};
+
+pub(crate) struct PtxRunner;
+
+impl PtxRunner {
+    pub fn spawn<'scope, 'view: 'scope>(
+        scope: &Scope<'scope>,
+        base_view: &'view (impl StateView + Sync),
+        finalizer: PtxFinalizerClient,
+    ) -> PtxRunnerClient {
+        let (work_tx, work_rx) = channel();
+        let client = PtxRunnerClient { work_tx };
+        scope.spawn(move |scope| {
+            let _timer = TIMER.timer_with(&["runner_block_total"]);
+
+            let first_command = work_rx.recv().expect("Channel closed.");
+            let mut manager = match first_command {
+                Command::SpawnWorkers {
+                    scheduler,
+                    num_workers,
+                } => WorkerManager::new(scope, num_workers, finalizer, base_view, scheduler),
+                _ => panic!("First command must be SpawnWorkers."),
+            };
+
+            loop {
+                match work_rx.recv().expect("Work thread died.") {
+                    Command::AddTransaction {
+                        txn_idx,
+                        transaction,
+                        dependencies,
+                    } => {
+                        manager.add_transaction(txn_idx, transaction, dependencies);
+                    },
+                    Command::FinishBlock => {
+                        manager.finish_block();
+                        break;
+                    },
+                    Command::SpawnWorkers { .. } => panic!("SpawnWorkers called twice."),
+                }
+            }
+        });
+        client
+    }
+}
+
+enum Command {
+    SpawnWorkers {
+        scheduler: PtxSchedulerClient,
+        num_workers: usize,
+    },
+    AddTransaction {
+        txn_idx: TxnIdx,
+        transaction: Transaction,
+        dependencies: HashMap<StateKey, Option<StateValue>>,
+    },
+    FinishBlock,
+}
+
+#[derive(Clone)]
+pub(crate) struct PtxRunnerClient {
+    work_tx: Sender<Command>,
+}
+
+impl PtxRunnerClient {
+    pub fn spawn_workers(&self, scheduler: PtxSchedulerClient, num_workers: usize) {
+        self.send_to_manager(Command::SpawnWorkers {
+            scheduler,
+            num_workers,
+        })
+    }
+
+    pub fn add_transaction(
+        &self,
+        txn_idx: TxnIdx,
+        transaction: Transaction,
+        dependencies: HashMap<StateKey, Option<StateValue>>,
+    ) {
+        self.send_to_manager(Command::AddTransaction {
+            txn_idx,
+            transaction,
+            dependencies,
+        })
+    }
+
+    /// Called by upstream to indicate all transactions form the block are added
+    pub fn finish_block(&self) {
+        self.send_to_manager(Command::FinishBlock)
+    }
+
+    fn send_to_manager(&self, command: Command) {
+        self.work_tx.send(command).expect("Manager died.")
+    }
+}
+
+type WorkerIndex = usize;
+
+struct WorkerManager {
+    finalizer: PtxFinalizerClient,
+    work_txs: Vec<Sender<WorkerCommand>>,
+    worker_ready_rx: Receiver<WorkerIndex>,
+}
+
+impl WorkerManager {
+    fn new<'scope, 'view: 'scope, BaseView: StateView + Sync>(
+        scope: &Scope<'scope>,
+        num_workers: usize,
+        finalizer: PtxFinalizerClient,
+        base_view: &'view BaseView,
+        scheduler: PtxSchedulerClient,
+    ) -> Self {
+        let (worker_ready_tx, worker_ready_rx) = channel();
+        let work_txs = (0..num_workers)
+            .map(|worker_idx| {
+                Worker::<BaseView>::spawn(
+                    scope,
+                    finalizer.clone(),
+                    worker_ready_tx.clone(),
+                    worker_idx,
+                    base_view,
+                    scheduler.clone(),
+                )
+            })
+            .collect();
+
+        Self {
+            finalizer,
+            work_txs,
+            worker_ready_rx,
+        }
+    }
+
+    fn add_transaction(
+        &mut self,
+        txn_idx: TxnIdx,
+        transaction: Transaction,
+        dependencies: HashMap<StateKey, Option<StateValue>>,
+    ) {
+        // wait for next available worker
+        let worker_index = {
+            let _timer = TIMER.timer_with(&["runner_wait_ready_worker"]);
+            self.worker_ready_rx.recv().expect("Channel closed.")
+        };
+        // send transaction to worker for execution
+        self.work_txs[worker_index]
+            .send(WorkerCommand::ExecuteTransaction {
+                txn_idx,
+                transaction,
+                dependencies,
+            })
+            .expect("Worker died.");
+    }
+
+    fn finish_block(&mut self) {
+        for work_tx in &self.work_txs {
+            work_tx.send(WorkerCommand::Finish).expect("Worker died.")
+        }
+        // Wait for all workers to quit (all txns finished).
+        while self.worker_ready_rx.recv().is_ok() {}
+
+        // Inform the finalizer the that block is finished, after all work has surely finished.
+        self.finalizer.finish_block();
+    }
+}
+
+enum WorkerCommand {
+    ExecuteTransaction {
+        txn_idx: TxnIdx,
+        transaction: Transaction,
+        dependencies: HashMap<StateKey, Option<StateValue>>,
+    },
+    Finish,
+}
+
+struct Worker<'view, BaseView> {
+    finalizer: PtxFinalizerClient,
+    work_rx: Receiver<WorkerCommand>,
+    worker_ready_tx: Sender<WorkerIndex>,
+    worker_index: WorkerIndex,
+    base_view: &'view BaseView,
+    scheduler: PtxSchedulerClient,
+}
+
+impl<'scope, 'view: 'scope, BaseView: StateView + Sync> Worker<'view, BaseView> {
+    fn spawn(
+        scope: &Scope<'scope>,
+        finalizer: PtxFinalizerClient,
+        worker_ready_tx: Sender<WorkerIndex>,
+        worker_index: WorkerIndex,
+        base_view: &'view BaseView,
+        scheduler: PtxSchedulerClient,
+    ) -> Sender<WorkerCommand> {
+        let (work_tx, work_rx) = channel();
+        scope.spawn(move |_scope| {
+            let worker = Self {
+                finalizer,
+                work_rx,
+                worker_ready_tx,
+                worker_index,
+                base_view,
+                scheduler,
+            };
+            worker.work()
+        });
+        work_tx
+    }
+
+    fn work(self) {
+        // Share a VM in the same thread.
+        // TODO(ptx): maybe warm up vm like done in AptosExecutorTask
+        let vm = AptosVM::new(&self.base_view.as_move_resolver());
+        let idx = format!("{}", self.worker_index);
+
+        // Inform the manger worker is up.
+        self.inform_worker_ready();
+
+        loop {
+            match self.work_rx.recv().expect("Sender died.") {
+                WorkerCommand::ExecuteTransaction {
+                    txn_idx,
+                    transaction,
+                    dependencies,
+                } => {
+                    let _total = PER_WORKER_TIMER.timer_with(&[&idx, "execute_txn_total"]);
+                    let _pre = PER_WORKER_TIMER.timer_with(&[&idx, "execute_txn_pre_vm"]);
+                    trace!("worker {} gonna run txn {}", self.worker_index, txn_idx);
+                    let state_view =
+                        OverlayedStateView::new_with_overlay(self.base_view, dependencies);
+                    let log_context = AdapterLogSchema::new(self.base_view.id(), txn_idx);
+                    let preprocessed_txn = preprocess_transaction::<AptosVM>(transaction);
+                    drop(_pre);
+                    let vm_output = {
+                        let _vm = PER_WORKER_TIMER.timer_with(&[&idx, "per_worker_vm"]);
+                        vm.execute_single_transaction(
+                            &preprocessed_txn,
+                            &vm.as_move_resolver(&state_view),
+                            &log_context,
+                        )
+                    };
+                    let _post = PER_WORKER_TIMER.timer_with(&[&idx, "per_worker_post_vm"]);
+                    // TODO(ptx): error handling
+                    let (_vm_status, vm_output, _msg) = vm_output.expect("VM execution failed.");
+
+                    // inform output state values to the manager
+                    for (key, op) in vm_output.change_set().write_set_iter() {
+                        self.scheduler
+                            .inform_state_value((key.clone(), txn_idx), op.as_state_value());
+                    }
+
+                    self.finalizer.add_vm_output(txn_idx, vm_output);
+                    self.inform_worker_ready();
+                    trace!("worker {} finished txn {}", self.worker_index, txn_idx);
+                },
+                WorkerCommand::Finish => {
+                    trace!("worker {} exit.", self.worker_index);
+                    break;
+                },
+            } // end match command
+        } // end loop
+    }
+
+    fn inform_worker_ready(&self) {
+        self.worker_ready_tx
+            .send(self.worker_index)
+            .expect("Manager died.");
+    }
+}
